+++ bootique-jdbc-hikaricp-instrumented/src/main/java/io/bootique/jdbc/instrumented/hikaricp/HikariCPMetricsInitializer.java
@@ -1,9 +1,9 @@
 package io.bootique.jdbc.instrumented.hikaricp;
 
-import com.codahale.metrics.Gauge;
 import com.codahale.metrics.MetricRegistry;
 import com.zaxxer.hikari.HikariDataSource;
 import io.bootique.jdbc.DataSourceListener;
+import io.bootique.jdbc.instrumented.hikaricp.metrics.HikariMetricsBridge;
 
 import javax.sql.DataSource;
 
@@ -12,22 +12,10 @@ import javax.sql.DataSource;
  */
 public class HikariCPMetricsInitializer implements DataSourceListener {
 
-    static final String METRIC_CATEGORY = "pool";
-    static final String METRIC_NAME_WAIT = "Wait";
-
-    static final String METRIC_NAME_TOTAL_CONNECTIONS = "TotalConnections";
-    static final String METRIC_NAME_IDLE_CONNECTIONS = "IdleConnections";
-    static final String METRIC_NAME_ACTIVE_CONNECTIONS = "ActiveConnections";
-    static final String METRIC_NAME_PENDING_CONNECTIONS = "PendingConnections";
-
     private MetricRegistry metricRegistry;
 
     public HikariCPMetricsInitializer(MetricRegistry metricRegistry) {
         this.metricRegistry = metricRegistry;
-    }
-
-    public static String waitTimerMetricName(String dataSourcePoolName) {
-        return MetricRegistry.name(dataSourcePoolName, METRIC_CATEGORY, METRIC_NAME_WAIT);
     }
 
     @Override
@@ -37,7 +25,7 @@ public class HikariCPMetricsInitializer implements DataSourceListener {
 
     @Override
     public void afterStartup(String name, String jdbcUrl, DataSource dataSource) {
-        collectPoolMetrics(name, dataSource);
+        collectPoolMetrics(dataSource);
     }
 
     @Override
@@ -45,40 +33,8 @@ public class HikariCPMetricsInitializer implements DataSourceListener {
         // do nothing
     }
 
-    void collectPoolMetrics(String name, DataSource dataSource) {
-
+    void collectPoolMetrics(DataSource dataSource) {
         HikariDataSource hikariDS = (HikariDataSource) dataSource;
-
-        addWaitTimer(hikariDS);
-        addTotalConnectionsGauge(hikariDS);
-        addIdleConnectionsGauge(hikariDS);
-        addActiveConnectionsGauge(hikariDS);
-        addPendingConnectionsGauge(hikariDS);
-    }
-
-
-    private void addWaitTimer(HikariDataSource ds) {
-        metricRegistry.timer(waitTimerMetricName(ds.getPoolName()));
-    }
-
-    private void addTotalConnectionsGauge(HikariDataSource ds) {
-        metricRegistry.register(MetricRegistry.name(ds.getPoolName(), METRIC_CATEGORY, METRIC_NAME_TOTAL_CONNECTIONS),
-                (Gauge<Integer>) () -> ds.getHikariPoolMXBean().getTotalConnections());
-    }
-
-    private void addIdleConnectionsGauge(HikariDataSource ds) {
-        metricRegistry.register(MetricRegistry.name(ds.getPoolName(), METRIC_CATEGORY, METRIC_NAME_IDLE_CONNECTIONS),
-                (Gauge<Integer>) () -> ds.getHikariPoolMXBean().getIdleConnections());
-    }
-
-    private void addActiveConnectionsGauge(HikariDataSource ds) {
-        metricRegistry.register(MetricRegistry.name(ds.getPoolName(), METRIC_CATEGORY, METRIC_NAME_ACTIVE_CONNECTIONS),
-                (Gauge<Integer>) () -> ds.getHikariPoolMXBean().getActiveConnections());
-    }
-
-    private void addPendingConnectionsGauge(HikariDataSource ds) {
-        metricRegistry.register(MetricRegistry.name(ds.getPoolName(), METRIC_CATEGORY, METRIC_NAME_PENDING_CONNECTIONS),
-                (Gauge<Integer>) () -> ds.getHikariPoolMXBean().getThreadsAwaitingConnection());
-
+        hikariDS.setMetricsTrackerFactory((pn, ps) -> new HikariMetricsBridge(pn, ps, metricRegistry));
     }
 }